# apps/backend/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/lookup/villas": {
      "get": {
        "operationId": "LookupController_getVillas",
        "summary": "Get list of villas (sites)",
        "description": "Returns a simplified list of all sites/villas for dropdown selection. Includes id, code, and name.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of villas retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "code": {
                        "type": "string",
                        "example": "VIL01"
                      },
                      "name": {
                        "type": "string",
                        "example": "Villa 1"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lookup"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/lookup/technicians": {
      "get": {
        "operationId": "LookupController_getTechnicians",
        "summary": "Get list of technicians",
        "description": "Returns a simplified list of active technicians for dropdown selection. Includes id, name, email, and initials.",
        "parameters": [
          {
            "name": "siteId",
            "required": false,
            "in": "query",
            "description": "Optional site ID to filter technicians by site",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of technicians retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "email": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string",
                        "nullable": true
                      },
                      "lastName": {
                        "type": "string",
                        "nullable": true
                      },
                      "fullName": {
                        "type": "string"
                      },
                      "initials": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lookup"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/lookup/villa-types": {
      "get": {
        "operationId": "LookupController_getVillaTypes",
        "summary": "Get list of villa types with default values",
        "description": "Returns all active villa type configurations with their default bedroom count, floor count, and area. Used for dropdowns and auto-filling villa forms.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of villa types retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "villaType": {
                        "type": "string",
                        "example": "1BHK"
                      },
                      "displayName": {
                        "type": "string",
                        "nullable": true,
                        "example": "1 Bedroom Hall Kitchen"
                      },
                      "defaultBedroomCount": {
                        "type": "number",
                        "nullable": true,
                        "example": 1
                      },
                      "defaultFloorCount": {
                        "type": "number",
                        "nullable": true,
                        "example": 1
                      },
                      "defaultAreaSqm": {
                        "type": "number",
                        "nullable": true,
                        "example": 50.5
                      },
                      "displayOrder": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lookup"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/lookup/cities": {
      "get": {
        "operationId": "LookupController_getCities",
        "summary": "Get list of cities",
        "description": "Returns all active cities for dropdown selection. Used in villa address forms.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of cities retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "example": "Dubai"
                      },
                      "code": {
                        "type": "string",
                        "nullable": true,
                        "example": "DXB"
                      },
                      "country": {
                        "type": "string",
                        "nullable": true,
                        "example": "UAE"
                      },
                      "region": {
                        "type": "string",
                        "nullable": true,
                        "example": "Dubai"
                      },
                      "locationDescription": {
                        "type": "string",
                        "nullable": true,
                        "example": "coast"
                      },
                      "displayOrder": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lookup"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/lookup/locations": {
      "get": {
        "operationId": "LookupController_getLocations",
        "summary": "Get list of locations",
        "description": "Returns all active locations (areas/neighborhoods) for dropdown selection. Can be filtered by cityId. Used in villa address forms.",
        "parameters": [
          {
            "name": "cityId",
            "required": false,
            "in": "query",
            "description": "Optional city ID to filter locations by city",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of locations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "cityId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "example": "Downtown Dubai"
                      },
                      "code": {
                        "type": "string",
                        "nullable": true,
                        "example": "DTD"
                      },
                      "country": {
                        "type": "string",
                        "nullable": true,
                        "example": "UAE"
                      },
                      "region": {
                        "type": "string",
                        "nullable": true,
                        "example": "Dubai"
                      },
                      "cityName": {
                        "type": "string",
                        "nullable": true,
                        "example": "Dubai"
                      },
                      "displayOrder": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "lookup"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/public/lookup/villas": {
      "get": {
        "operationId": "PublicLookupController_getPublicVillas",
        "summary": "Public: Get list of villas (sites) without authentication",
        "description": "Returns a simplified list of all sites/villas for pre-login flows like site selection. Includes id, code, and name.",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of villas retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "code": {
                        "type": "string",
                        "example": "VIL01"
                      },
                      "name": {
                        "type": "string",
                        "example": "Villa 1"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "public-lookup"
        ]
      }
    },
    "/api/v1/villas": {
      "get": {
        "operationId": "VillaController_findAll",
        "summary": "Get all villas",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Villas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "VillaController_create",
        "summary": "Create a new villa",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Villas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TENX API",
    "description": "Multi-tenant SaaS ERP backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
...
  }
}